feat:eventService

+++ gateway/src/main/java/ru/practicum/explorewithme/publ/PublicController.java
@@ -51,7 +51,7 @@ public class PublicController {
     }
 
     @GetMapping("/events/{eventId}")
-    public ResponseEntity<Object> getEventById(@PathVariable(name = "eventId") int eventId) {
+    public ResponseEntity<Object> getEventById(@PathVariable(name = "eventId") long eventId) {
         return client.getEventById(eventId);
     }
